+++ src/Client/ClientCore/Systems/ClientState/ClientStateController.cs
@@ -15,6 +15,8 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 using Sovereign.ClientCore.Events.Details;
+using Sovereign.ClientCore.Rendering.Scenes.Game.World;
+using Sovereign.EngineCore.Components.Types;
 using Sovereign.EngineCore.Events;
 using Sovereign.EngineCore.Events.Details;
 
@@ -23,7 +25,7 @@ namespace Sovereign.ClientCore.Systems.ClientState;
 /// <summary>
 ///     Public API for the ClientState system.
 /// </summary>
-public class ClientStateController
+public class ClientStateController(HighlightManager highlightManager)
 {
     /// <summary>
     ///     Sets a state flag.
@@ -76,5 +78,52 @@ public class ClientStateController
     public void DeselectItem(IEventSender eventSender)
     {
         SelectItem(eventSender, -1);
+    }
+
+    /// <summary>
+    ///     Synchronously clears any block highlights.
+    /// </summary>
+    public void ClearBlockHighlights()
+    {
+        highlightManager.ClearBlockHighlight();
+    }
+
+    /// <summary>
+    ///     Synchronously adds a block highlight.
+    /// </summary>
+    /// <param name="position">Position.</param>
+    public void AddBlockHighlight(GridPosition position)
+    {
+        highlightManager.AddBlockHighlight(position);
+    }
+
+    /// <summary>
+    ///     Synchronously adds block highlights to a rectangle in the XY plane.
+    /// </summary>
+    /// <param name="bottomLeft">Bottom left corner of rectangle.</param>
+    /// <param name="width">Width in blocks.</param>
+    /// <param name="height">Height in blocks.</param>
+    public void AddBlockHighlightXyRect(GridPosition bottomLeft, uint width, uint height)
+    {
+        highlightManager.AddBlockHighlightXyRect(bottomLeft, width, height);
+    }
+
+    /// <summary>
+    ///     Synchronously adds block highlights to a square centered on the given position.
+    /// </summary>
+    /// <param name="center">Center position.</param>
+    /// <param name="width">Width of the full square in blocks.</param>
+    public void AddBlockHighlightSquare(GridPosition center, uint width)
+    {
+        highlightManager.AddBlockHighlightSquare(center, width);
+    }
+
+    /// <summary>
+    ///     Synchronously sets the block highlight state.
+    /// </summary>
+    /// <param name="state"></param>
+    public void SetBlockHighlightState(Highlight state)
+    {
+        highlightManager.BlockHighlightState = state;
     }
 }